Refactor call feature components and fix bugs
